MES-394 - Refactor counter records and active time calculation

+++ src/main/java/com/alcegory/mescloud/model/dto/PlcMqttDto.java
@@ -10,7 +10,7 @@ public class PlcMqttDto extends AbstractMqttDto {
     private String productionOrderCode;
     private String equipmentCode;
     private int equipmentStatus;
-    private long activeTime;
+    private int activeTime;
     private int[] alarms;
     private CounterMqttDto[] counters;
 }
